Move back to a single assets generator
